Combined all dispersion tests in a single test file
